django_ninja/requirements.txt
faker==30.8.2
Django==5.1.2
django-ninja==1.3.0
orjson==3.10.11
psycopg[binary]==3.2.3
